refactored commands into api so it can be used programmatically. Started with init :boom:
