+++ src/api/index.js
@@ -1,0 +1,1 @@
+export { defautl as init } from './init';
