+++ src/index.js
@@ -1,2 +1,5 @@
+import * as api from './api';
+
 export { default as cli } from './cli';
 export { default as parser } from './parser';
+export default api;
